// apps/ecommerce-demo/src/types/glasses.ts
export interface Glasses {
  id: number;
  name: string;
  description: string;
  price: string;
  thumbnail: string;
  model_3d_file: string;
  scale_factor: number;
  offset_x: number;
  offset_y: number;
  offset_z: number;
  rotation_x?: number;
  rotation_y?: number;
  rotation_z?: number;
  created_at: string;
  updated_at: string;
}